Move JsonModelConverterInterface and JsonTagInfoConverterInterface to separate files
